third_party/blink/web_tests/webmidi/requestmidiaccess-basic-expected.txt

CONSOLE WARNING: line 12: Web MIDI will ask a permission to use even if the sysex is not specified in the MIDIOptions since M77, around September 2019. See https://www.chromestatus.com/feature/5138066234671104 for more details.
Test if requestMIDIAccess is available.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS navigator.requestMIDIAccess is defined.
PASS promise is defined.
PASS promise.then is defined.
PASS successfullyParsed is true

TEST COMPLETE
